+++ tests/conftest.py
@@ -13,8 +13,13 @@ pytest_plugins = ("mex.common.testing.plugin",)
 TEST_USER_DATABASE = {
     APIKey("api-key-one"): ["test_system"],
     APIKey("john-doe"): ["test_system", "foo_system"],
-    APIKey("api-test-key"): ["test_system", "foo_system"],
+    APIKey("api-test-key"): [
+        "test_system",
+        "foo_system",
+        "x_system_that_does_not_exist",
+    ],
     APIKey("alice"): ["foo_system"],
+    APIKey("i-do-what-i-want"): ["admin"],
 }
 
 
@@ -23,7 +28,7 @@ def settings(tmp_path: Path) -> DropSettings:
     """Load the settings for this pytest session."""
     settings = DropSettings.get()
     settings.drop_directory = str(tmp_path)
-    settings.drop_user_database = TEST_USER_DATABASE
+    settings.drop_api_key_database = TEST_USER_DATABASE
     return settings