+++ tests/test_main.py
@@ -1,4 +1,6 @@
+import json
 from pathlib import Path
+from typing import Any
 from unittest.mock import AsyncMock, call
 
 import pytest
@@ -8,6 +10,25 @@ from pytest import MonkeyPatch
 import mex
 from mex.drop.settings import DropSettings
 from mex.drop.types import EntityType, XSystem
+
+
+@pytest.fixture
+def dropped_data(tmp_path: Path) -> dict[str, Any]:
+    """Dictionary of api key, x-system name, entity type and content."""
+    settings = DropSettings.get()
+    settings.drop_directory = tmp_path
+    data = {
+        "content": {"foo": "bar"},
+        "api_key": "api-test-key",
+        "x_system": "test_system",
+        "entity_type": "foo",
+    }
+    x_system_dir = tmp_path / data["x_system"]
+    expected_file = x_system_dir / f"{data['entity_type']}.json"
+    x_system_dir.mkdir(parents=True)
+    with expected_file.open("w") as handle:
+        json.dump(data["content"], handle)
+    return data
 
 
 @pytest.mark.parametrize(
@@ -60,13 +81,38 @@ def test_drop_data(
         assert mocked_sink.call_args == call(expected_content, expected_file)
 
 
-def test_show_form(client: TestClient) -> None:
+def test_download_data(client: TestClient, dropped_data: dict[str, Any]) -> None:
+    client.headers.update({"X-API-Key": dropped_data["api_key"]})
     response = client.get(
-        "/v0/test_system/test_type",
+        f"/v0/{dropped_data['x_system']}/{dropped_data['entity_type']}",
     )
+    assert response.status_code == 200, response
+    assert response.json() == dropped_data["content"]
 
+    response = client.get(
+        f"/v0/{dropped_data['x_system']}/entity_that_does_not_exist",
+    )
+    assert response.status_code == 404, response
+
+
+def test_list_x_systems(client: TestClient, dropped_data: dict[str, Any]) -> None:
+    response = client.get("/v0/", headers={"X-API-Key": "i-do-what-i-want"})
     assert response.status_code == 200, response.text
-    assert "<title>mex-drop</title>" in response.text
+    assert response.json() == {"x-systems": [f"{dropped_data['x_system']}"]}
+
+
+def test_list_entity_types(client: TestClient, dropped_data: dict[str, Any]) -> None:
+    client.headers.update({"X-API-Key": dropped_data["api_key"]})
+    response = client.get(
+        f"/v0/{dropped_data['x_system']}",
+    )
+    assert response.status_code == 200, response.text
+    assert response.json() == {"entity-types": [f"{dropped_data['entity_type']}"]}
+
+    response = client.get(
+        "/v0/x_system_that_does_not_exist",
+    )
+    assert response.status_code == 404, response.text
 
 
 def test_health_check(client: TestClient) -> None: