+++ CHANGELOG.md
@@ -16,14 +16,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cruft template link
 - open-code workflow
 - add helm charts for k8s deployments
+- endpoint for listing available x-systems (requires authentication for x-system `admin`)
+- endpoint for listing downloadable files for an x-system
+- endpoint for downloading data
 
 ### Changes
 
 - harmonized boilerplate
+- updated dependencies
+- **BREAKING** rename setting `MEX_DROP_USER_DATABASE` to `MEX_DROP_API_KEY_DATABASE`
 
 ### Deprecated
 
 ### Removed
+
+- **BREAKING** remove the upload html form at `/v0/{x_system}/{entity_type}`, because the endpoint conflicted with the most obvious route for downloading files (GET /v0/x-system/entity)
 
 ### Fixed